+++ src/test/java/de/is24/deadcode4j/analyzer/hibernateannotations/ClassDefiningGenericGenerator.java
@@ -1,0 +1,7 @@
+package de.is24.deadcode4j.analyzer.hibernateannotations;
+
+import org.hibernate.annotations.GenericGenerator;
+
+@GenericGenerator(name = "aGenerator", strategy = "IndependentClass")
+public class ClassDefiningGenericGenerator {
+}